fix(sync): delegate platform sync to package runtime
- Add a shared package locator for installed agent-infra runtimes
- Move the full GitHub platform adapter into package-owned runtime code
- Keep downstream templates as thin delegates without project npm installs
- Cover init, update, validation, and Windows command lookup paths
