+++ package.json
@@ -26,6 +26,7 @@
     "!dist/**/*.map",
     "bin/cli.ts",
     "lib/",
+    "runtime/",
     "templates/"
   ],
   "engines": {
@@ -45,6 +46,7 @@
   "dependencies": {
     "@clack/prompts": "1.7.0",
     "@larksuiteoapi/node-sdk": "^1.68.0",
+    "cross-spawn": "^7.0.6",
     "picocolors": "1.1.1",
     "semver": "^7.8.1",
     "smol-toml": "^1.6.1",
@@ -63,6 +65,7 @@
     "prepublishOnly": "npm run build && node --experimental-strip-types --no-warnings --test \"tests/**/*.test.ts\""
   },
   "devDependencies": {
+    "@types/cross-spawn": "^6.0.6",
     "@types/node": "^22.20.0",
     "@types/semver": "^7.7.1",
     "typescript": "~7.0"
