+++ tests/unit/core/release.test.ts
@@ -72,11 +72,13 @@ test("package metadata supports scoped npm publishing", () => {
     "!dist/**/*.map",
     "bin/cli.ts",
     "lib/",
+    "runtime/",
     "templates/"
   ]);
   assert.deepEqual(Object.keys(pkg.dependencies).sort(), [
     "@clack/prompts",
     "@larksuiteoapi/node-sdk",
+    "cross-spawn",
     "picocolors",
     "semver",
     "smol-toml",
@@ -85,6 +87,7 @@ test("package metadata supports scoped npm publishing", () => {
   assert.match(pkg.scripts.prepublishOnly, /npm run build/);
   assert.match(pkg.scripts.prepublishOnly, /--test/);
   assert.match(pkg.scripts.prepublishOnly, /tests\/\*\*\/\*\.test\.ts/);
+  assert.match(read("runtime/platform-adapters/platform-sync.github.js"), /from "cross-spawn"/);
 });
 
 test("Node runtime baseline and release publisher baseline stay pinned", () => {
